#28 Lazy Location Computation
Store the offset and the RiddlParserInput in the
Location object and defer computation of the line and column until
they are needed. Consequently, successful compilations won't be
penalized by precomputing error message locations

This involved passing the RiddlParserInput in many places and especially
in all the test cases.

Although the original ticket, #28 wanted to replace the Location object
with just an integer offset, this is not viable without the context of
the input because otherwise the line/col cannot be computed at all.